Move fetcher functions to separate module
